Fix remaining VC++ warnings of types C4244 (http://msdn.microsoft.com/en-us/library/th7a07tz(VS.80).aspx), C4804 (http://msdn.microsoft.com/en-us/library/b6x9ad5e.aspx) and C4146 (http://msdn.microsoft.com/en-us/library/4kh09110.aspx). Also pick off a few of the remaining warnings. With this checkin, only four warning types are left at W3: C4101, C4267, C4800 and C4805
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@8563 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/serial/Midi_w32.cc
@@ -82,8 +82,8 @@ static char inlongmes[OPENMSX_W32_MIDI_SYSMES_MAXLEN];
 
 static void w32_midiDevNameConv(char *dst, char *src)
 {
-	unsigned len = strlen(src);
-	unsigned i;
+	size_t len = strlen(src);
+	size_t i;
 	for (i = 0; i < len; ++i) {
 		if ((src[i] < '0') || (src[i] > 'z') ||
 		    ((src[i] > '9') && (src[i] < 'A')) ||
